move the changes to the operator module
