+++ core/src/main/java/org/keycloak/representations/idm/ComponentExportRepresentation.java
@@ -18,7 +18,6 @@
 package org.keycloak.representations.idm;
 
 import org.keycloak.common.util.MultivaluedHashMap;
-import io.fabric8.crd.generator.annotation.SchemaFrom;
 
 /**
  * @author <a href="mailto:mposolda@redhat.com">Marek Posolda</a>
@@ -29,9 +28,7 @@ public class ComponentExportRepresentation {
     private String name;
     private String providerId;
     private String subType;
-    @SchemaFrom(type = org.keycloak.representations.overrides.NoSubcomponentsComponentExportRepresentationMap.class)
     private MultivaluedHashMap<String, ComponentExportRepresentation> subComponents = new MultivaluedHashMap<>();
-    @SchemaFrom(type = org.keycloak.representations.overrides.MultivaluedStringStringHashMap.class)
     private MultivaluedHashMap<String, String> config = new MultivaluedHashMap<>();
 
     public String getId() {
