+++ operator/src/main/java/org/keycloak/operator/v2alpha1/crds/RealmSpec.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 Red Hat, Inc. and/or its affiliates
+ * Copyright 2021 Red Hat, Inc. and/or its affiliates
  * and other contributors as indicated by the @author tags.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -14,13 +14,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.keycloak.crd;
+package org.keycloak.operator.v2alpha1.crds;
 
 import org.keycloak.representations.idm.RealmRepresentation;
 
 import javax.validation.constraints.NotNull;
 
-public class ExampleRealmCRDSpec {
+public class RealmSpec {
 
     @NotNull
     private String keycloakCRName;
@@ -42,4 +42,5 @@ public class ExampleRealmCRDSpec {
     public void setRealm(RealmRepresentation realm) {
         this.realm = realm;
     }
+
 }
