+++ operator/src/main/java/org/keycloak/representations/idm/ComponentRepresentation.java
@@ -18,6 +18,7 @@
 package org.keycloak.representations.idm;
 
 import org.keycloak.common.util.MultivaluedHashMap;
+import io.fabric8.crd.generator.annotation.SchemaFrom;
 
 /**
  * @author <a href="mailto:mposolda@redhat.com">Marek Posolda</a>
@@ -32,6 +33,7 @@ public class ComponentRepresentation {
     private String providerType;
     private String parentId;
     private String subType;
+    @SchemaFrom(type = org.keycloak.representations.overrides.MultivaluedStringStringHashMap.class)
     private MultivaluedHashMap<String, String> config;
 
     public String getId() {
