+++ operator/src/main/java/org/keycloak/representations/overrides/NoSubGroupsGroupRepresentation.java
