+++ operator/src/main/java/org/keycloak/representations/overrides/NoSubcomponentsComponentExportRepresentation.java
